server/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altgraph==0.17.3
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
buildozer==1.5.0
CacheControl==0.13.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
cssbeautifier==1.14.8
distlib==0.3.7
Django==4.2.2
django-cors-headers==4.1.0
django-decouple==2.1
django-filter==23.3
django-rest-framework==0.1.0
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docutils==0.20.1
docx2pdf==0.1.8
drf-spectacular==0.26.3
drf-spectacular-sidecar==2023.7.1
EditorConfig==0.12.3
Faker==18.10.1
filelock==3.12.2
fire==0.5.0
fonttools==4.42.1
frozenlist==1.4.0
gunicorn==21.2.0
html5lib==1.1
idna==3.4
inflection==0.5.1
instaloader==4.10.1
jsbeautifier==1.14.8
jsonschema==4.17.3
msgpack==1.0.5
multidict==6.0.4
openai==0.27.8
opencv-python==4.8.0.76
pandas==2.0.3
passlib==1.7.4
pdf2docx==0.5.6
pdfkit==1.0.0
pefile==2023.2.7
pexpect==4.8.0
Pillow==9.5.0
pipdeptree==2.13.0
platformdirs==3.10.0
psycopg2==2.9.9
psycopg2-binary==2.9.6
ptyprocess==0.7.0
Pygments==2.16.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.7
PyJWT==2.8.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-docx==0.8.11
pytz==2023.3
pywin32-ctypes==0.2.2
PyYAML==6.0
requests==2.31.0
sh==2.0.6
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
termcolor==2.3.0
toposort==1.10
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
virtualenv==20.24.3
webencodings==0.5.1
whitenoise==6.6.0
yarl==1.9.2
